refactor: preferences reducer
