+++ src/component/reducer/preferences/panelsPreferencesDefaultValues.ts
