+++ src/component/reducer/preferences/utilites/filterObject.ts
@@ -1,0 +1,7 @@
+export function filterObject(data: any) {
+  return JSON.parse(JSON.stringify(data), (key, value) => {
+    if (value?.hidden !== true) {
+      return value;
+    }
+  });
+}
