+++ src/component/reducer/preferences/utilites/flatObject.ts
@@ -1,0 +1,9 @@
+export function flatObject(data: any) {
+  let result = {};
+  JSON.parse(JSON.stringify(data), (key, value) => {
+    if (value?.hidden !== true && key) {
+      result[key] = result[key]++ || 1;
+    }
+  });
+  return result;
+}
