+++ src/component/reducer/preferences/utilites/getActiveWorkspace.ts
@@ -1,0 +1,7 @@
+import { Draft } from 'immer';
+
+import { PreferencesState } from '../preferencesReducer';
+
+export function getActiveWorkspace(draft: Draft<PreferencesState>) {
+  return draft.workspaces[draft.workspace.current || 'default'];
+}
