+++ src/component/reducer/preferences/utilites/getPreferencesByWorkspace.ts
@@ -1,0 +1,6 @@
+import { NMRiumWorkspace } from '../../../NMRium';
+import Workspaces from '../../../workspaces';
+
+export function getPreferencesByWorkspace(workspace: NMRiumWorkspace) {
+  return Workspaces?.[workspace] || {};
+}
